+++ native/homm5-editor.c
@@ -79,6 +79,7 @@
 #include "qol/fix-encourage.c"
 #include "qol/fix-barbarian-learning.c"
 #include "qol/fix-snare-crash.c"
+#include "qol/fix-payback.c"
 
 BOOL WINAPI DllMain(HINSTANCE self, DWORD reason, LPVOID reserved) {
   (void)reserved;
@@ -139,6 +140,7 @@ BOOL WINAPI DllMain(HINSTANCE self, DWORD reason, LPVOID reserved) {
   if (g_qol[QOL_ENCOURAGE_FIX]) install_encourage_fix();
   if (g_qol[QOL_BARBARIAN_LEARNING_FIX]) install_barbarian_learning_fix();
   if (g_qol[QOL_SNARE_CRASH_FIX]) install_snare_fix();
+  if (g_qol[QOL_PAYBACK_FIX]) install_payback_fix();
   return TRUE;
 }